[KYUUBI #1757] Extract workingDir from engine specific process builder into ProcBuilder
