+++ kyuubi-server/src/main/scala/org/apache/kyuubi/engine/flink/FlinkProcessBuilder.scala
@@ -99,32 +99,6 @@ class FlinkProcessBuilder(
 
   override protected def commands: Array[String] = Array(executable)
 
-  override protected val workingDir: Path = {
-    env.get("KYUUBI_WORK_DIR_ROOT").map { root =>
-      val workingRoot = Paths.get(root).toAbsolutePath
-      if (!Files.exists(workingRoot)) {
-        debug(s"Creating KYUUBI_WORK_DIR_ROOT at $workingRoot")
-        Files.createDirectories(workingRoot)
-      }
-      if (Files.isDirectory(workingRoot)) {
-        workingRoot.toString
-      } else null
-    }.map { rootAbs =>
-      val working = Paths.get(rootAbs, proxyUser)
-      if (!Files.exists(working)) {
-        debug(s"Creating $proxyUser's working directory at $working")
-        Files.createDirectories(working)
-      }
-      if (Files.isDirectory(working)) {
-        working
-      } else {
-        Utils.createTempDir(rootAbs, proxyUser)
-      }
-    }.getOrElse {
-      Utils.createTempDir(namePrefix = proxyUser)
-    }
-  }
-
   override def toString: String = commands.map {
     case arg if arg.startsWith("--") => s"\\\n\t$arg"
     case arg => arg
